修复一个新增为0触发事件的问题

+++ spring-domain-event/src/main/java/com/gitee/spring/domain/event/repository/EventRepository.java
@@ -30,70 +30,82 @@ public class EventRepository extends ConfiguredRepository {
     @Override
     public int insert(BoundedContext boundedContext, Object entity) {
         int count = super.insert(boundedContext, entity);
-        RepositoryEvent repositoryEvent = new RepositoryEvent(this);
-        repositoryEvent.setMethodName("insert");
-        repositoryEvent.setOperationType(OperationType.INSERT);
-        repositoryEvent.setBoundedContext(boundedContext);
-        repositoryEvent.setEntity(entity);
-        applicationContext.publishEvent(repositoryEvent);
+        if (count != 0) {
+            RepositoryEvent repositoryEvent = new RepositoryEvent(this);
+            repositoryEvent.setMethodName("insert");
+            repositoryEvent.setOperationType(OperationType.INSERT);
+            repositoryEvent.setBoundedContext(boundedContext);
+            repositoryEvent.setEntity(entity);
+            applicationContext.publishEvent(repositoryEvent);
+        }
         return count;
     }
 
     @Override
     public int update(BoundedContext boundedContext, Object entity) {
         int count = super.update(boundedContext, entity);
-        RepositoryEvent repositoryEvent = new RepositoryEvent(this);
-        repositoryEvent.setMethodName("update");
-        repositoryEvent.setOperationType(OperationType.UPDATE);
-        repositoryEvent.setBoundedContext(boundedContext);
-        repositoryEvent.setEntity(entity);
-        applicationContext.publishEvent(repositoryEvent);
+        if (count != 0) {
+            RepositoryEvent repositoryEvent = new RepositoryEvent(this);
+            repositoryEvent.setMethodName("update");
+            repositoryEvent.setOperationType(OperationType.UPDATE);
+            repositoryEvent.setBoundedContext(boundedContext);
+            repositoryEvent.setEntity(entity);
+            applicationContext.publishEvent(repositoryEvent);
+        }
         return count;
     }
 
     @Override
     public int updateByExample(Object entity, Object example) {
         int count = super.updateByExample(entity, example);
-        RepositoryEvent repositoryEvent = new RepositoryEvent(this);
-        repositoryEvent.setMethodName("updateByExample");
-        repositoryEvent.setOperationType(OperationType.UPDATE);
-        repositoryEvent.setEntity(entity);
-        repositoryEvent.setExample(example);
-        applicationContext.publishEvent(repositoryEvent);
+        if (count != 0) {
+            RepositoryEvent repositoryEvent = new RepositoryEvent(this);
+            repositoryEvent.setMethodName("updateByExample");
+            repositoryEvent.setOperationType(OperationType.UPDATE);
+            repositoryEvent.setEntity(entity);
+            repositoryEvent.setExample(example);
+            applicationContext.publishEvent(repositoryEvent);
+        }
         return count;
     }
 
     @Override
     public int delete(BoundedContext boundedContext, Object entity) {
         int count = super.delete(boundedContext, entity);
-        RepositoryEvent repositoryEvent = new RepositoryEvent(this);
-        repositoryEvent.setMethodName("delete");
-        repositoryEvent.setOperationType(OperationType.DELETE);
-        repositoryEvent.setBoundedContext(boundedContext);
-        repositoryEvent.setEntity(entity);
-        applicationContext.publishEvent(repositoryEvent);
+        if (count != 0) {
+            RepositoryEvent repositoryEvent = new RepositoryEvent(this);
+            repositoryEvent.setMethodName("delete");
+            repositoryEvent.setOperationType(OperationType.DELETE);
+            repositoryEvent.setBoundedContext(boundedContext);
+            repositoryEvent.setEntity(entity);
+            applicationContext.publishEvent(repositoryEvent);
+        }
         return count;
     }
 
     @Override
     public int deleteByPrimaryKey(Object primaryKey) {
         int count = super.deleteByPrimaryKey(primaryKey);
-        RepositoryEvent repositoryEvent = new RepositoryEvent(this);
-        repositoryEvent.setMethodName("deleteByPrimaryKey");
-        repositoryEvent.setOperationType(OperationType.DELETE);
-        repositoryEvent.setPrimaryKey(primaryKey);
-        applicationContext.publishEvent(repositoryEvent);
+        if (count != 0) {
+            RepositoryEvent repositoryEvent = new RepositoryEvent(this);
+            repositoryEvent.setMethodName("deleteByPrimaryKey");
+            repositoryEvent.setOperationType(OperationType.DELETE);
+            repositoryEvent.setPrimaryKey(primaryKey);
+            applicationContext.publishEvent(repositoryEvent);
+        }
         return count;
     }
 
     @Override
     public int deleteByExample(Object example) {
         int count = super.deleteByExample(example);
-        RepositoryEvent repositoryEvent = new RepositoryEvent(this);
-        repositoryEvent.setMethodName("deleteByExample");
-        repositoryEvent.setOperationType(OperationType.DELETE);
-        repositoryEvent.setExample(example);
-        applicationContext.publishEvent(repositoryEvent);
+        if (count != 0) {
+            RepositoryEvent repositoryEvent = new RepositoryEvent(this);
+            repositoryEvent.setMethodName("deleteByExample");
+            repositoryEvent.setOperationType(OperationType.DELETE);
+            repositoryEvent.setExample(example);
+            applicationContext.publishEvent(repositoryEvent);
+        }
         return count;
     }
